refactor sdata testutil
